% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVec.R
\name{checkVec}
\alias{checkVec}
\title{Subsidiary Power Calculation Function}
\usage{
checkVec(val)
}
\arguments{
\item{val}{parameter}
}
\value{
Either a string message or val.
}
\description{
Check if input is of length greater than 1 and convert to string message if so.
}
\note{
This function is called internally and thus should not be used directly.
}
\author{
Kirk Bansak and Eddie Yang
}
