% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_powerLATE.R
\name{print.powerLATE}
\alias{print.powerLATE}
\title{Print Function for powerLATE}
\usage{
\method{print}{powerLATE}(message.input, message.output, res, note)
}
\arguments{
\item{message.input}{vector of strings specifying input parameters.}

\item{message.output}{string prompt specifying the target parameter.}

\item{res}{dataframe of output parameters.}

\item{note}{note specifying assumptions made and additional details.}
}
\value{
strings and a dataframe for output.
}
\description{
Print output for powerLATE and powerLATE.cov.
}
\note{
This function is called internally and should not be used directly.
}
\author{
Kirk Bansak and Eddie Yang
}
