% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{.onAttach}
\alias{.onAttach}
\title{Subsidiary PowerLATE Function}
\usage{
.onAttach(lib, pkg)
}
\arguments{
\item{lib}{libname}

\item{pkg}{package name}
}
\description{
Subsidiary function to perform power calculation under equal assignment probability and ordered mean assumption.
}
\references{
Bansak, K. (2020). A Generalized Approach to Power Analysis for Local Average Treatment Effects. Statistical Science, 35(2), 254-271.
}
