% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerLATE_cov.R
\name{powerLATE.cov}
\alias{powerLATE.cov}
\title{A Generalized Approach to Power Analysis for Local Average Treatment Effects (with covariates)}
\usage{
powerLATE.cov(pZ = 0.5, pi, N, kappa,
	tau = NULL, sig.level = 0.05, power,
	effect.size = TRUE, omega = NULL, assume.ord.means = FALSE, r2dw, r2yw)
}
\arguments{
\item{pZ}{probability of being assigned to treatment. Default is 0.5, i.e. equal assignment probability.}

\item{pi}{compliance rate. Equivalently, average causal effect of Z on D.}

\item{N}{total number of observations}

\item{kappa}{effect size}

\item{tau}{absolute effect. Must be supplied if effect.size==FALSE.}

\item{sig.level}{significance level (Type I error probability). Default is 0.05.}

\item{power}{power of test (1 minus Type II error probability)}

\item{effect.size}{whether effect size rather than tau is used in subsequent calculations. Default is TRUE.}

\item{omega}{pooled standard deviation. Must be supplied if effect.size==FALSE.}

\item{assume.ord.means}{whether ordered mean assumption should be made. Default is FALSE}

\item{r2dw}{proportion of variation in D left unexplained by Z that is explained by W.}

\item{r2yw}{proportion of variation in Y left unexplained by Z that is explained by W.}
}
\value{
A dataframe with lower bounds on the target parameter, along with supplied parameter values. If one of \{kappa, N, power, pi, tau\} is passed as a vector of values, a dataframe with no. of rows = length of the vector will be returned.
}
\description{
Main function to perform generalized power analysis for LATE with covariates
}
\details{
Exactly two of the parameters \{kappa, N, power\} must be supplied, from which the third (target) parameter will be calculated. If effect.size==FALSE, exactly two of the parameters \{tau, N, power\} must be supplied.
}
\examples{

#######################################################
### Example 1: 
### LATE w/ equal assigment w/o ordered mean assumption
#######################################################

res <- powerLATE.cov(pZ=0.5, pi=0.35, kappa=seq(0.4, 1.0, 0.1), power=0.8,
                     r2dw=0.2, r2yw=0.15)

# also returns an invisible object: output.parameter
res$output.parameter

#######################################################
### Example 2: 
### LATE w/o equal assigment w/ ordered mean assumption
#######################################################

res <- powerLATE.cov(pZ=0.67, pi=0.35, kappa=seq(0.4, 1.0, 0.1), 
                power=0.8, assume.ord.means=TRUE, r2dw=0.2, r2yw=0.15)

# also returns an invisible object: output.parameter
res$output.parameter
}
\references{
Bansak, K. (2020). A Generalized Approach to Power Analysis for Local Average Treatment Effects. Statistical Science, 35(2), 254-271.
}
\author{
Kirk Bansak
}
