% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerLATE.R
\name{powerLATE}
\alias{powerLATE}
\title{A Generalized Approach to Power Analysis for Local Average Treatment Effects}
\usage{
powerLATE(pZ = 0.5, pi, N, kappa, 
	tau = NULL, sig.level = 0.05, power, 
	effect.size = TRUE, omega = NULL, assume.ord.means = FALSE)
}
\arguments{
\item{pZ}{probability of being assigned to treatment. Default is 0.5, i.e. equal assignment probability.}

\item{pi}{compliance rate. Equivalently, average causal effect of Z on D.}

\item{N}{total number of observations}

\item{kappa}{effect size}

\item{tau}{absolute effect. Must be supplied if effect.size==FALSE.}

\item{sig.level}{significance level (Type I error probability). Default is 0.05.}

\item{power}{power of test (1 minus Type II error probability)}

\item{effect.size}{whether effect size rather than tau is used in subsequent calculations. Default is TRUE.}

\item{omega}{pooled standard deviation. Must be supplied if effect.size==FALSE.}

\item{assume.ord.means}{whether ordered mean assumption should be made. Default is FALSE}
}
\value{
A dataframe with lower bounds on the target parameter, along with supplied parameter values. If one of \{kappa, N, power, pi, tau\} is passed as a vector of values, a dataframe with no. of rows = length of the vector will be returned.
}
\description{
Main function to perform generalized power analysis for LATE
}
\details{
Exactly two of the parameters \{kappa, N, power\} must be supplied, from which the third (target) parameter will be calculated. If effect.size==FALSE, exactly two of the parameters \{tau, N, power\} must be supplied.
}
\references{
Bansak, K. (2020). A Generalized Approach to Power Analysis for Local Average Treatment Effects. Statistical Science, 35(2), 254-271.
}
\author{
Kirk Bansak
}
