% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerLATE.R
\name{powerLATE}
\alias{powerLATE}
\title{A Generalized Approach to Power Analysis for Local Average Treatment Effects}
\usage{
powerLATE(pZ = 0.5, pi, N, kappa,
	sig.level = 0.05, power,
	effect.size = TRUE, tau = NULL, omega = NULL,
	assume.ord.means = FALSE)
}
\arguments{
\item{pZ}{probability of being assigned to treatment. Default is 0.5, i.e. equal assignment probability.}

\item{pi}{compliance rate. Equivalently, average causal effect of treatment assignment on treatment uptake.}

\item{N}{total sample size.}

\item{kappa}{effect size.}

\item{sig.level}{significance level (Type I error probability). Default is 0.05.}

\item{power}{power of test (1 minus Type II error probability).}

\item{effect.size}{whether effect size (kappa) rather than absolute effect (tau) is used in power calculations. Default is \code{TRUE}.}

\item{tau}{absolute effect. Must only be supplied if \code{effect.size = FALSE}.}

\item{omega}{within-group standard deviation of the outcome. Must be supplied if \code{effect.size = FALSE}. See Details.}

\item{assume.ord.means}{whether ordered means assumption is made. Default is \code{FALSE}. See Details.}
}
\value{
A list that includes the values of the input parameters supplied by the user (\code{input.parameter}) and the corresponding output value(s) 
of the target parameter (\code{output.parameter}).

Note also that the results along with additional information will be displayed in the console.
}
\description{
Function to perform generalized power analysis for the LATE (i.e. under noncompliance with treatment assignment). 
Function allows for user to work with either standardized effect sizes or absolute effects. 
The results provided presume a test of the null hypothesis that the LATE equals 0 with a two-sided alternative.
}
\details{
If \code{effect.size = TRUE} (the default setting), exactly two of the parameters \{\code{kappa, N, power}\} must be supplied, 
from which the third (target) parameter will be calculated. If \code{effect.size = FALSE}, \code{omega} must be supplied, and exactly two of 
the parameters \{\code{tau, N, power}\} must be supplied. \code{pi} must always be supplied, and the user can change \code{pZ} and \code{sig.level} 
from their default values.

The user may also supply one of \{\code{kappa, N, power, pi, tau}\} as a vector of values to perform multiple power calculations at a time, 
in which case the target parameter will be calculated for that entire vector.

If \code{effect.size = FALSE}, \code{omega} represents the reference within-assignment-group standard deviation of the outcome. The user may wish to 
use an estimate of the standard deviation of the outcome prior to the intervention (i.e. in the absence of the treatment). 
See "Discussion on Effect Sizes" section in Bansak (2020) for more information and guidance.

The \code{assume.ord.means} argument allows the user to choose whether or not to make the ordered means assumption, presented and described in Bansak (2020). 
Users should only make this assumption (i.e. set \code{assume.ord.means = TRUE}) if they are reasonably confident that it will be met in their context of interest. 
See "Narrowing the Bounds" section in Bansak (2020) for more information and guidance.
}
\examples{

#######################################################
### Example 1: 
### LATE w/ equal assigment w/o ordered mean assumption
#######################################################

res <- powerLATE(pZ=0.5, pi=0.35, kappa=seq(0.4, 1.0, 0.1), power=0.8)

# also returns an invisible object: output.parameter
res$output.parameter

#######################################################
### Example 2: 
### LATE w/o equal assigment w/ ordered mean assumption
#######################################################

res <- powerLATE(pZ=0.67, pi=0.35, kappa=seq(0.4, 1.0, 0.1), 
                power=0.8, assume.ord.means=TRUE)

# also returns an invisible object: output.parameter
res$output.parameter
}
\references{
Bansak, K. (2020). A Generalized Approach to Power Analysis for Local Average Treatment Effects. Statistical Science, 35(2), 254-271.
}
\author{
Kirk Bansak and Eddie Yang
}
