% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_powerLATE.R
\name{print.powerLATE}
\alias{print.powerLATE}
\title{Print Function for powerLATE}
\usage{
\method{print}{powerLATE}(x, ...)
}
\arguments{
\item{x}{List of message.input, message.output, res, note to be printed}

\item{...}{Further arguments to be passed to \code{print.powerLATE()}.}
}
\value{
strings and a dataframe for output.
}
\description{
Print output for powerLATE and powerLATE.cov.
}
\note{
This function is called internally and should not be used directly.
}
\author{
Kirk Bansak and Eddie Yang
}
